+++ src/main/java/com/example/SpringPlayground/githubWebhook/model/github/Commit.java
@@ -1,6 +1,6 @@
 package com.example.SpringPlayground.githubWebhook.model.github;
 
-import com.example.SpringPlayground.githubWebhook.model.slack.blocks.Text;
+import com.example.SpringPlayground.githubWebhook.model.slack.component.Text;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
