+++ src/main/java/com/example/SpringPlayground/githubWebhook/model/github/GithubUser.java
@@ -1,8 +1,8 @@
 package com.example.SpringPlayground.githubWebhook.model.github;
 
-import com.example.SpringPlayground.githubWebhook.model.slack.blocks.Text;
-import com.example.SpringPlayground.githubWebhook.model.slack.blocks.image.Image;
-import com.example.SpringPlayground.githubWebhook.model.slack.blocks.image.TextImage;
+import com.example.SpringPlayground.githubWebhook.model.slack.component.Text;
+import com.example.SpringPlayground.githubWebhook.model.slack.component.image.Image;
+import com.example.SpringPlayground.githubWebhook.model.slack.component.image.TextImage;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
