+++ src/main/java/com/example/SpringPlayground/githubWebhook/model/github/RepositoryInfo.java
@@ -1,7 +1,5 @@
 package com.example.SpringPlayground.githubWebhook.model.github;
 
-import com.example.SpringPlayground.githubWebhook.model.slack.blocks.Section;
-import com.example.SpringPlayground.githubWebhook.model.slack.blocks.Text;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
