+++ src/main/java/com/example/SpringPlayground/githubWebhook/model/github/ping/Hook.java
@@ -1,7 +1,7 @@
 package com.example.SpringPlayground.githubWebhook.model.github.ping;
 
-import com.example.SpringPlayground.githubWebhook.model.slack.blocks.Section;
-import com.example.SpringPlayground.githubWebhook.model.slack.blocks.Text;
+import com.example.SpringPlayground.githubWebhook.model.slack.component.Section;
+import com.example.SpringPlayground.githubWebhook.model.slack.component.Text;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
