+++ src/main/java/com/example/SpringPlayground/githubWebhook/model/github/ping/PingMessage.java
@@ -3,9 +3,9 @@ package com.example.SpringPlayground.githubWebhook.model.github.ping;
 import com.example.SpringPlayground.githubWebhook.model.github.RepositoryInfo;
 import com.example.SpringPlayground.githubWebhook.model.github.GithubMessage;
 import com.example.SpringPlayground.githubWebhook.model.slack.SlackPayload;
-import com.example.SpringPlayground.githubWebhook.model.slack.blocks.Divider;
-import com.example.SpringPlayground.githubWebhook.model.slack.blocks.Header;
-import com.example.SpringPlayground.githubWebhook.model.slack.blocks.Section;
+import com.example.SpringPlayground.githubWebhook.model.slack.component.Divider;
+import com.example.SpringPlayground.githubWebhook.model.slack.component.Header;
+import com.example.SpringPlayground.githubWebhook.model.slack.component.Section;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
