+++ src/main/java/com/example/SpringPlayground/githubWebhook/model/github/pullrequest/PRMessage.java
@@ -4,10 +4,10 @@ import com.example.SpringPlayground.githubWebhook.model.github.GithubUser;
 import com.example.SpringPlayground.githubWebhook.model.github.RepositoryInfo;
 import com.example.SpringPlayground.githubWebhook.model.github.GithubMessage;
 import com.example.SpringPlayground.githubWebhook.model.slack.SlackPayload;
-import com.example.SpringPlayground.githubWebhook.model.slack.blocks.Divider;
-import com.example.SpringPlayground.githubWebhook.model.slack.blocks.Header;
-import com.example.SpringPlayground.githubWebhook.model.slack.blocks.Section;
-import com.example.SpringPlayground.githubWebhook.model.slack.blocks.image.TextImage;
+import com.example.SpringPlayground.githubWebhook.model.slack.component.Divider;
+import com.example.SpringPlayground.githubWebhook.model.slack.component.Header;
+import com.example.SpringPlayground.githubWebhook.model.slack.component.Section;
+import com.example.SpringPlayground.githubWebhook.model.slack.component.image.TextImage;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
