+++ src/main/java/com/example/SpringPlayground/githubWebhook/model/github/push/PushMessage.java
@@ -5,8 +5,8 @@ import com.example.SpringPlayground.githubWebhook.model.github.GithubUser;
 import com.example.SpringPlayground.githubWebhook.model.github.RepositoryInfo;
 import com.example.SpringPlayground.githubWebhook.model.github.GithubMessage;
 import com.example.SpringPlayground.githubWebhook.model.slack.SlackPayload;
-import com.example.SpringPlayground.githubWebhook.model.slack.blocks.*;
-import com.example.SpringPlayground.githubWebhook.model.slack.blocks.image.TextImage;
+import com.example.SpringPlayground.githubWebhook.model.slack.component.*;
+import com.example.SpringPlayground.githubWebhook.model.slack.component.image.TextImage;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
