--- src/main/java/com/example/SpringPlayground/githubWebhook/model/slack/blocks/SlackPayloadComponent.java
@@ -1,5 +1,0 @@
-package com.example.SpringPlayground.githubWebhook.model.slack.blocks;
-
-public interface SlackPayloadComponent {
-
-}
