+++ src/main/java/com/example/SpringPlayground/githubWebhook/model/slack/component/SlackPayloadComponent.java
@@ -1,0 +1,5 @@
+package com.example.SpringPlayground.githubWebhook.model.slack.component;
+
+public interface SlackPayloadComponent {
+
+}
