+++ src/main/java/com/example/SpringPlayground/githubWebhook/model/slack/component/Text.java
@@ -1,4 +1,4 @@
-package com.example.SpringPlayground.githubWebhook.model.slack.blocks;
+package com.example.SpringPlayground.githubWebhook.model.slack.component;
 
 import lombok.Data;
