+++ src/test/java/com/example/SpringPlayground/githubWebhook/service/SlackMessageTest.java
@@ -2,10 +2,10 @@ package com.example.SpringPlayground.githubWebhook.service;
 
 import com.example.SpringPlayground.githubWebhook.constants.Constants;
 import com.example.SpringPlayground.githubWebhook.model.slack.SlackPayload;
-import com.example.SpringPlayground.githubWebhook.model.slack.blocks.Section;
-import com.example.SpringPlayground.githubWebhook.model.slack.blocks.Text;
-import com.example.SpringPlayground.githubWebhook.model.slack.blocks.image.Image;
-import com.example.SpringPlayground.githubWebhook.model.slack.blocks.image.TextImage;
+import com.example.SpringPlayground.githubWebhook.model.slack.component.Section;
+import com.example.SpringPlayground.githubWebhook.model.slack.component.Text;
+import com.example.SpringPlayground.githubWebhook.model.slack.component.image.Image;
+import com.example.SpringPlayground.githubWebhook.model.slack.component.image.TextImage;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.slack.api.Slack;
